# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Flipic API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "localhost:4101",
  "basePath": "/",
  "tags": [
    {
      "name": "Quiz",
      "description": "Endpoints"
    }
  ],
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "apiKeyAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "X-API-KEY",
      "description": "any description..."
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/quiz": {
      "post": {
        "tags": [
          "Quiz"
        ],
        "description": "Endpoint to create a quiz",
        "parameters": [
          {
            "name": "quiz",
            "in": "body",
            "description": "Quiz information.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Quiz"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      },
      "put": {
        "tags": [
          "Quiz"
        ],
        "description": "Endpoint to edit a quiz",
        "parameters": [
          {
            "name": "quiz",
            "in": "body",
            "description": "Quiz information.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Quiz"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/quiz/{id}": {
      "get": {
        "tags": [
          "Quiz"
        ],
        "description": "Endpoint to get a quiz",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/quiz-all": {
      "get": {
        "tags": [
          "Quiz"
        ],
        "description": "Endpoint to get all quiz",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    }
  },
  "definitions": {
    "Item": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Item name"
        },
        "icon": {
          "type": "string",
          "example": "Icon name"
        }
      }
    },
    "Quiz": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Quiz Name"
        },
        "hash": {
          "type": "string",
          "example": "Quiz Hash"
        },
        "items": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Item"
          }
        }
      }
    }
  }
}